aihero/course/pyproject.toml
[project]
name = "course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.68.0",
    "pydantic-ai>=0.0.15",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "tqdm>=4.67.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
